my_pytorch_project/requirements.txt
torch
torchvision
torchaudio
matplotlib
pyyaml
